Volume name should respect restrictions
In foreman, the vm name contains also the domain name.
When setting this name to the volume name, running the VM fails
with the following error:

  Warning  FailedCreate  21s               virtualmachine-controller
  Error creating pod: Pod "virt-launcher-vmi-multus-lqcwz" is invalid:
  spec.containers[0].name: Invalid value: "volumevmi-multus.example.com":
  a DNS-1123 label must consist of lower case alphanumeric characters or '-',
  and must start and end with an alphanumeric character (e.g. 'my-name',
  or '123-abc', regex used for validation is '[a-z0-9]([-a-z0-9]*[a-z0-9])?')

Therefore the dots from FQDN vm name cannot be part of the volume name.
The patch assumes the name without the domain should be sufficient.