Replace VmVolume with Volume as main collection
The 'volumes' collection will allow to list all volumes in the system
that are backed by a PVC or to list all of vm volumes, if given a vm
name while trying to list the volumes.

The Volume class will be used also as a parameter for passing volume for
a virtual machine create request, so additional information could be
specified such as 'bus' and 'boot_order' to determine how the volume
should be configured for the vm.

With this PR also fixed the delete_vm service so cleanup after tests
will work as expected when recording VCR for specs.