Fix undefined method exception handling 404 error
If an API call raises a `Fog::Kubevirt::Errors::ClientError` exception
due to a 404 the `def all` methods are catching that exception and
setting the result to `[]`.

Later on the method then tries to set the `@kind` and
`@resource_version` instance variables to e.g. `vms.kind` and
`vms.resource_version` which fails if `vms = []`.